5th_Nov/project/requirements.txt
fastapi
uvicorn
streamlit
langchain
chromadb
openai
tiktoken
pypdf
tqdm
python-dotenv
